big reorg and cleanup
