+++ globus/__init__.py
@@ -1,0 +1,5 @@
+import logging as _logging
+
+logger = _logging.getLogger(__name__)
+logger.setLevel(_logging.DEBUG)
+logger.addHandler(_logging.NullHandler())
