+++ globus/uilts.py
@@ -1,0 +1,5 @@
+import sys
+
+
+def is_interactive():
+    return sys.stdin.isatty()
