+++ setup.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from setuptools import setup
+from setuptools import setup, find_packages
 
 THIS_DIR = Path(__file__).parent
 
@@ -14,7 +14,7 @@ setup(
     long_description=Path("README.md").read_text(),
     long_description_content_type="text/markdown",
     url="https://github.com/JoshKarpel/globus-transfer",
-    py_modules=["globus"],
-    entry_points={"console_scripts": ["globus = globus:cli"]},
+    py_modules=find_packages(),
+    entry_points={"console_scripts": ["globus = globus.cli:cli"]},
     install_requires=Path("requirements.txt").read_text().splitlines(),
 )
